requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
matplotlib==1.3.1
nose==1.3.1
numpy==1.8.1
openpyxl==1.6.2
pandas==0.13.1
pyparsing==2.0.2
python-dateutil==2.2
pytz==2014.2
scipy==0.13.3
six==1.6.1
tornado==3.2
wsgiref==0.1.2
xlrd==0.9.2
xlwt==0.7.5
